Adds admin form component, move filtered colums to mixin
