+++ app/components/admin-form.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-admin/components/admin-form';
