+++ app/templates/admin/model-records/show/edit.hbs
@@ -1,2 +1,38 @@
 model-record/show/edit
-{{outlet}}
+
+<div class='edit'>
+  <h2>Edit {{model.id}}</h2>
+
+  {{admin-form
+    recordType=recordType
+    model=model
+    save=(action 'save')
+    delete=(action 'delete')
+    cancel=(action 'cancel')
+    onUpdate=(action 'onFieldUpdate')
+  }}
+
+  <div class='relationships'>
+    <h2>Relationships</h2>
+
+    {{#each belongsTo as |relationship|}}
+      <h4 data-test='belongs-to-{{relationship.name}}'>{{relationship.name}}</h4>
+      {{admin-table
+        recordType=relationship.type
+        records=relationship.records}}
+    {{/each}}
+
+    {{#each hasMany as |relationship|}}
+      <h4 data-test='has-many-{{relationship.name}}'>
+        {{#link-to
+          'admin.model-records.show.related'
+          recordType
+          model.id
+          relationship.name
+          }}
+          {{relationship.name}}
+        {{/link-to}}
+      </h4>
+    {{/each}}
+  </div>
+</div>
